+++ src/Geomancy.hs
@@ -1,3 +1,5 @@
+{-# LANGUAGE PatternSynonyms #-}
+
 module Geomancy
   ( Mat4
 
@@ -10,20 +12,23 @@ module Geomancy
   , Vec2
   , vec2
   , withVec2
+  , pattern WithVec2
 
   , Vec3
   , vec3
   , withVec3
+  , pattern WithVec3
 
   , Vec4
   , vec4
   , withVec4
+  , pattern WithVec4
 
   ) where
 
 import Geomancy.Mat4 (Mat4)
 import Geomancy.Transform (Transform(..))
 import Geomancy.Quaternion (Quaternion, quaternion, withQuaternion)
-import Geomancy.Vec2 (Vec2, vec2, withVec2)
-import Geomancy.Vec3 (Vec3, vec3, withVec3)
-import Geomancy.Vec4 (Vec4, vec4, withVec4)
+import Geomancy.Vec2 (Vec2, vec2, withVec2, pattern WithVec2)
+import Geomancy.Vec3 (Vec3, vec3, withVec3, pattern WithVec3)
+import Geomancy.Vec4 (Vec4, vec4, withVec4, pattern WithVec4)